Final commit
Final commit, created a variable that that stated the title, then moved on to  for loop that was for the race. I created a function that contained 2 loops nested inside of eachother. This function contained all the information on the candidates. Finally, I made a last function to display the percentage bars for the candidates using a switch/case to identify the different parties.